social sign in service test added. code refactored.
